Stop waiting for creation when a backup gets cancelled
Since 1.27, the backups can get cancelled. This implies that
a new status CANCELLED exists. In the case of a backup creation
which is waiting for it to finished, it won't stop if the status
is different that SUCCEDED or FAILED. This commit adds a condition
for the CANCELED status during creation.